+++ auctions_ai/tests/factories.py
@@ -8,4 +8,3 @@ class UserFactory(factory.django.DjangoModelFactory):
 
     username = factory.Sequence(lambda n: "user%d" % n)
     email = factory.LazyAttribute(lambda obj: "%s@example.com" % obj.username)
-    password = factory.PostGenerationMethodCall("set_password", "defaultpassword")
